+++ src/commands/npmRelease.js
@@ -27,6 +27,7 @@ program
     .name('taoRelease npmRelease')
     .usage('[options]')
     .option(...cliOptions.debug)
+    .option(...cliOptions.releaseVersion)
     // options with defaults
     .option(...cliOptions.baseBranch)
     .option(...cliOptions.branchPrefix)
@@ -53,13 +54,14 @@ async function npmRelease() {
         await release.verifyLocalChanges();
         await release.signTags();
         await release.verifyBranches();
+        await release.extractVersion();
         await release.doesTagExists();
         await release.doesReleasingBranchExists();
         await release.isReleaseRequired();
         await release.confirmRelease();
         await release.createReleasingBranch();
+        await release.updateVersion();
         await release.initialiseGithubClient();
-        await release.pushReleasingBranch();
         await release.createPullRequest();
         await release.extractReleaseNotes();
         await release.mergePullRequest();