+++ src/commands/oldWayRelease.js
@@ -27,6 +27,7 @@ program
     .name('taoRelease oldWayRelease')
     .usage('[options]')
     .option(...cliOptions.debug)
+    .option(...cliOptions.releaseVersion)
     // options with defaults
     .option(...cliOptions.baseBranch)
     .option(...cliOptions.branchPrefix)
@@ -58,12 +59,14 @@ async function releaseExtension() {
         await release.verifyLocalChanges();
         await release.signTags();
         await release.verifyBranches();
+        await release.extractVersion();
         await release.doesTagExists();
         await release.doesReleasingBranchExists();
         await release.isReleaseRequired();
         await release.confirmRelease();
         await release.createReleasingBranch();
         await release.build();
+        await release.updateVersion();
         await release.initialiseGithubClient();
         await release.createPullRequest();
         await release.extractReleaseNotes();