+++ src/commands/prepareRelease.js
@@ -27,6 +27,7 @@ program
     .name('taoRelease prepareRelease')
     .usage('[options]')
     .option(...cliOptions.debug)
+    .option(...cliOptions.releaseVersion)
     // options with defaults
     .option(...cliOptions.baseBranch)
     .option(...cliOptions.branchPrefix)
@@ -55,11 +56,13 @@ async function prepareRelease() {
         await release.initialiseGitClient();
         await release.verifyLocalChanges();
         await release.verifyBranches();
+        await release.extractVersion();
         await release.doesTagExists();
         await release.doesReleasingBranchExists();
         await release.isReleaseRequired();
         await release.createReleasingBranch();
         await release.build();
+        await release.updateVersion();
 
         log.done('Release branch prepared, and pushed to remote.').exit();
     } catch (error) {