+++ HISTORY.md
@@ -48,3 +48,9 @@
 ## [Version 1.1.1]
 
  - Bugfix for when data.extension is not already set
+
+ ## [Version 1.2.0]
+
+ - Calculate the version from conventional commits. More details https://www.conventionalcommits.org/en/v1.0.0/
+ - Introduce new CLI parameter `release-version` which provide version to be taken as new release version
+ - Update unit tests